perf(format): shrink ParseEvent from 96 to 48 bytes
Optimize core event types for better memory efficiency:

- Span: 16 → 8 bytes (use u32 for offset/len instead of usize)
- FieldKey: 72 → 24 bytes (split into Name/Full variants, box metadata)
- ParseEventKind: 80 → 32 bytes
- ParseEvent: 96 → 48 bytes (50% reduction)

FieldKey now uses an enum with two variants:
- Name(Cow<str>): common case for JSON/YAML/TOML (24 bytes)
- Full(Box<FullFieldKey>): for Styx with doc/tag metadata (8 bytes)

This reduces stack frame sizes and improves cache utilization
without measurable performance regression on benchmarks.

+++ facet-toml/tests/integration/spanned.rs
@@ -54,7 +54,7 @@ fn spanned_string() {
 
     // Span should be populated and point to the value
     let span = config.name.span.expect("span should be populated");
-    let spanned_text = &input[span.offset..span.offset + span.len];
+    let spanned_text = &input[span.offset as usize..span.offset as usize + span.len as usize];
     assert_eq!(spanned_text, r#""foo""#);
 }
 
@@ -86,7 +86,7 @@ fn spanned_bool() {
     assert!(config.enabled.value);
 
     let span = config.enabled.span.expect("span should be populated");
-    let spanned_text = &input[span.offset..span.offset + span.len];
+    let spanned_text = &input[span.offset as usize..span.offset as usize + span.len as usize];
     assert_eq!(spanned_text, "true");
 }
 
@@ -102,7 +102,7 @@ fn spanned_integer() {
     assert_eq!(config.version.value, 42);
 
     let span = config.version.span.expect("span should be populated");
-    let spanned_text = &input[span.offset..span.offset + span.len];
+    let spanned_text = &input[span.offset as usize..span.offset as usize + span.len as usize];
     assert_eq!(spanned_text, "42");
 }
